latest rhizome

+++ rx4rdf/rx/rhizome/site-template.xsl
@@ -6,18 +6,31 @@
         xmlns:rdf='http://www.w3.org/1999/02/22-rdf-syntax-ns#'
         xmlns:wf='http://rx4rdf.sf.net/ns/racoon/xpath-ext#'
         xmlns:f = 'http://xmlns.4suite.org/ext'
-        exclude-result-prefixes = "f wf a wiki rdf" >		
+        xmlns:response-header='http://rx4rdf.sf.net/ns/racoon/http-response-header#'
+        xmlns:previous = 'http://rx4rdf.sf.net/ns/racoon/previous#'
+        xmlns:session = 'http://rx4rdf.sf.net/ns/racoon/session#'
+        exclude-result-prefixes = "f wf a wiki rdf response-header previous session" >		
     <xsl:param name="_contents" />		
-    <xsl:param name="title" />		
+    <xsl:param name="previous:title" />		
     <xsl:param name="_name" />		
     <xsl:param name="_prevnode" />		    
+    <xsl:param name="session:login" />
+    <xsl:param name="session:message" />
+    <xsl:param name="_url" />		 
+    <xsl:param name="response-header:content-type"/>
+    
     <xsl:output method='html' indent='no' />
+    
 <xsl:template match="/">
-<!-- html template based on http://www.projectseven.com/tutorials/css_t/example.htm -->
-<html lang="en">
+<!-- this page is always html, not the content's mimetype -->
+<xsl:variable name='content-type' select="wf:assign-metadata('response-header:content-type', 'text/html')" />
+<xsl:variable name='title' select="f:if($previous:title, $previous:title, f:if($_prevnode/wiki:title, $_prevnode/wiki:title, $_name) )" />
+<!-- html template based on http://www.projectseven.com/tutorials/css_t/example.htm 
+    (well, not much anymore, gave up and started using nested tables)
+-->
+<html>
 <head>
-<title><xsl:value-of select="f:if($title, $title, $_name)" /></title>
-<meta http-equiv="Content-Type" content="text/html; charset=iso-8859-1" />
+<title><xsl:value-of select="$title" /></title>
 <link href="basestyles.css" rel="stylesheet" type="text/css" />
 <link rel="icon" href="favicon.ico" />
 </head>
@@ -28,9 +41,29 @@
 <tr>
 <!-- Header -->
 <td id="header" width="100%" colspan="2">
-    <p>
-<a href="index"><img border="0" src="underconstruction.gif" /></a> Header, site title goes here: edit the<a href="site-template?action=edit">site template</a>
-    </p>
+<div style='float: right'> 
+         <xsl:choose>
+            <xsl:when test="$session:login">
+Welcome <xsl:value-of select="$session:login" />
+<form action='logout' method='POST' >
+<input TYPE="hidden" NAME="redirect" value="{$_url}" />
+<input type="submit" value="logout" name="logout"/>    
+</form>            
+            </xsl:when>
+            <xsl:otherwise>
+Login&#xa0;
+<form action='login' method='POST' >
+<input TYPE="text" NAME="loginname" />
+<input TYPE="password" NAME="password" />
+<input TYPE="hidden" NAME="redirect" value="{$_url}" />
+<input type="submit" value="login" name="login"/>    
+</form>            
+<br />Or <a href="users-guest?action=new">signup</a>
+           </xsl:otherwise>
+        </xsl:choose>
+<xsl:value-of select="$session:message" disable-output-escaping='yes' />
+</div>    
+<a href="index"><img border="0" src="{/*[wiki:name='site-template']/wiki:header-image}" /></a><xsl:value-of disable-output-escaping='yes' select="/*[wiki:name='site-template']/wiki:header-text" />
 </td>
 </tr>
 
@@ -58,7 +91,7 @@
     <!-- title bar -->
     <tr>
     <td valign="top" height='1' id="titlebar">
-    <xsl:value-of select="f:if($title, $title, $_name)" />
+    <xsl:value-of select="$title" />
     </td>
     </tr>
 
@@ -66,7 +99,7 @@
     <tr>
     <td valign="top" id="maincontent">
          <xsl:choose>
-            <xsl:when test="$_prevnode/a:contents/a:ContentTransform/a:transformed-by = 'http://rx4rdf.sf.net/ns/wiki#item-format-text'">
+            <xsl:when test="$response-header:content-type='text/plain' or $_prevnode/a:contents/a:ContentTransform/a:transformed-by = 'http://rx4rdf.sf.net/ns/wiki#item-format-text'">
             <pre>
             <xsl:value-of disable-output-escaping='no' select="$_contents" />
             </pre>
@@ -88,7 +121,7 @@
 <div style='float: right'><a href="edit">New</a>&#xa0;<a href="list?type=all">List</a></div>
 
 <a href="{$_name}">View</a>&#xa0;<a href="{$_name}?action=edit">Edit</a>&#xa0;<a href="{$_name}?action=showrevisions">Revisions</a>
-&#xa0;<a href="{$_name}?action=edit-metadata">Metadata</a>&#xa0;<a href="{$_name}?action=delete">Delete</a>
+&#xa0;<a href="{$_name}?action=edit-metadata">Metadata</a>&#xa0;<a href="{$_name}?action=confirm-delete">Delete</a>
 </p><p>       
 <form action='search'>
 
@@ -99,8 +132,8 @@ Type<select name="searchType">
     <option value="regex" >Regex</option>
     </select>
 &#xa0;View<select name="view">
-    <option value="compact" selected='selected'>Compact</option>
-    <option value="long" >Long</option>
+    <option value="html" selected='selected'>HTML</option>
+    <option value="rss91" >RSS .91</option>
     </select>
 &#xa0;<input type="submit" value="search" name="Search"/>
